Finished the flow of a session and added the recording of the camera.

+++ lib/Helpers/cameraHelpers.dart
@@ -11,33 +11,43 @@ Future<int> createCameraWithSettings(
         CameraDescription camera, MediaSettings settings) async =>
     await CameraPlatform.instance.createCameraWithSettings(camera, settings);
 
-Future<bool> anyCamerasAvailable() async =>
-    (await fetchCameras()).isNotEmpty;
+Future<bool> anyCamerasAvailable() async => (await fetchCameras()).isNotEmpty;
 
-Future<void> disposeCamera(int cameraId) async =>
-    await CameraPlatform.instance.dispose(cameraId);
+Future<void> disposeCamera(int cameraIdToDispose) async {
+  if (cameraIdToDispose == -1) return;
 
-Future<int> startRecording(
-    {int index = cameraIndex, MediaSettings settings = cameraSettings, Duration maxDuration = maxVideoDuration}) async {
+  await CameraPlatform.instance.dispose(cameraIdToDispose);
+  cameraId = -1;
+}
+
+Future<int> startRecording({
+  int index = cameraIndex,
+  MediaSettings settings = cameraSettings,
+  Duration maxDuration = maxVideoDuration,
+  bool doesDisposeOldCamera = true,
+}) async {
+  if (doesDisposeOldCamera) await disposeCamera(cameraId);
+
   List<CameraDescription> cameras = await fetchCameras();
   cameraId = await createCameraWithSettings(cameras[index], settings);
-  await CameraPlatform.instance.startVideoRecording(cameraId,maxVideoDuration: maxVideoDuration);
+  await CameraPlatform.instance
+      .startVideoRecording(cameraId, maxVideoDuration: maxVideoDuration);
   return cameraId;
 }
 
 Future<int> stopRecording({String? storeLocation}) async {
+  if (cameraId == -1) return cameraId;
+
   XFile videoFile = await CameraPlatform.instance.stopVideoRecording(cameraId);
   await disposeCamera(cameraId);
 
   // Checking if the file needs to be moved to a different location
   if (storeLocation != null) {
     final String orginFilePath = videoFile.path;
- 
+
     await videoFile.saveTo(storeLocation);
     deleteFile(orginFilePath);
   }
 
   return cameraId;
 }
-
-// stopRecording(storeLocation: 'Data/Sessions/0e75b8d8-aad5-4530-9f0e-d7a9c3bef69e/video.mp4');
